refactor: update concatStrings import path to common utility
