+++ frontend/src/community/people/components/organisms/SupervisorReassignmentModal/SupervisorReassignmentModal.tsx
@@ -21,7 +21,7 @@ import {
   SupervisorReassignmentActionType,
   TransferSupervisorsPayload
 } from "~community/people/types/PeopleTypes";
-import { concatStrings } from "~community/people/utils/jobFamilyUtils/commonUtils";
+import { concatStrings } from "~community/common/utils/commonUtil";
 
 interface SupervisorReassignmentModalProps {
   isOpen: boolean;
